# backend/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "GET IT NOW API",
    "version": "1.0.0",
    "description": "On-demand service marketplace API for connecting customers with verified workers",
    "contact": {
      "name": "API Support",
      "email": "support@getitnow.example"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Development server"
    },
    {
      "url": "https://api.getitnow.example",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/welfare/workers/me": {
      "get": {
        "summary": "Get worker welfare bundle",
        "tags": [
          "Welfare"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Worker welfare bundle"
          },
          "404": {
            "description": "Worker profile not found"
          }
        }
      }
    },
    "/welfare/workers/me/training": {
      "post": {
        "summary": "Add training record",
        "tags": [
          "Welfare"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "courseName"
                ],
                "properties": {
                  "courseName": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 200
                  },
                  "provider": {
                    "type": "string",
                    "maxLength": 200
                  },
                  "completedOn": {
                    "type": "string",
                    "format": "date"
                  },
                  "expiresOn": {
                    "type": "string",
                    "format": "date"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "planned",
                      "in_progress",
                      "completed",
                      "expired"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Training record created"
          }
        }
      }
    },
    "/welfare/workers/me/insurance": {
      "post": {
        "summary": "Add insurance record",
        "tags": [
          "Welfare"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "provider",
                  "policyReference",
                  "coverageAmount",
                  "startsOn",
                  "expiresOn"
                ],
                "properties": {
                  "provider": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 200
                  },
                  "policyReference": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 200
                  },
                  "coverageAmount": {
                    "type": "number",
                    "minimum": 0,
                    "maximum": 100000000
                  },
                  "startsOn": {
                    "type": "string",
                    "format": "date"
                  },
                  "expiresOn": {
                    "type": "string",
                    "format": "date"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "pending",
                      "active",
                      "expired",
                      "cancelled"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Insurance record created"
          }
        }
      }
    },
    "/welfare/workers/me/payout-account": {
      "put": {
        "summary": "Update payout account",
        "tags": [
          "Welfare"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "provider",
                  "accountReference"
                ],
                "properties": {
                  "provider": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 50
                  },
                  "accountReference": {
                    "type": "string",
                    "minLength": 4,
                    "maxLength": 200
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Payout account updated"
          }
        }
      }
    },
    "/welfare/workers/me/safety-incidents": {
      "get": {
        "summary": "Get safety incidents",
        "tags": [
          "Welfare"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of safety incidents"
          }
        }
      },
      "post": {
        "summary": "Report a safety incident",
        "tags": [
          "Welfare"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "type",
                  "severity",
                  "description"
                ],
                "properties": {
                  "bookingId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "injury",
                      "near_miss",
                      "equipment_failure",
                      "hazardous_condition",
                      "other"
                    ]
                  },
                  "severity": {
                    "type": "string",
                    "enum": [
                      "low",
                      "medium",
                      "high",
                      "critical"
                    ]
                  },
                  "description": {
                    "type": "string",
                    "minLength": 10,
                    "maxLength": 2000
                  },
                  "latitude": {
                    "type": "number",
                    "minimum": -90,
                    "maximum": 90
                  },
                  "longitude": {
                    "type": "number",
                    "minimum": -180,
                    "maximum": 180
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Safety incident reported"
          }
        }
      }
    },
    "/welfare/workers/me/benefits": {
      "get": {
        "summary": "Get worker benefits",
        "tags": [
          "Welfare"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of benefits and eligibility"
          }
        }
      }
    },
    "/welfare/workers/me/eligibility": {
      "get": {
        "summary": "Get benefit eligibility",
        "tags": [
          "Welfare"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Benefit eligibility status"
          }
        }
      }
    },
    "/welfare/benefits": {
      "get": {
        "summary": "List all benefits (admin)",
        "tags": [
          "Welfare"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of benefits"
          }
        }
      },
      "post": {
        "summary": "Create benefit (admin)",
        "tags": [
          "Welfare"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 200
                  },
                  "description": {
                    "type": "string",
                    "maxLength": 1000
                  },
                  "eligibilityCriteria": {
                    "type": "object"
                  },
                  "value": {
                    "type": "number",
                    "minimum": 0,
                    "maximum": 100000000
                  },
                  "provider": {
                    "type": "string",
                    "maxLength": 200
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "active",
                      "inactive"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Benefit created"
          }
        }
      }
    },
    "/welfare/benefits/{id}/eligibility": {
      "post": {
        "summary": "Set benefit eligibility for a worker (admin)",
        "tags": [
          "Welfare"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "workerId"
                ],
                "properties": {
                  "workerId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "eligible": {
                    "type": "boolean"
                  },
                  "expiresAt": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "metadata": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Eligibility set"
          }
        }
      }
    },
    "/welfare/workers/{workerId}/safety-incidents": {
      "get": {
        "summary": "Get safety incidents for a worker (admin)",
        "tags": [
          "Welfare"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "workerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of safety incidents"
          }
        }
      }
    },
...
  }
}